+++ src/main/kotlin/com/sexysisters/tojserverv2/domain/approve/ApproveReader.kt
@@ -1,6 +1,6 @@
 package com.sexysisters.tojserverv2.domain.approve
 
-import com.sexysisters.tojserverv2.domain.student.Student
+import com.sexysisters.tojserverv2.domain.student.domain.Student
 
 interface ApproveReader {
     fun getApprove(applicant: Student, acceptor: Student): Approve
