+++ src/main/kotlin/com/sexysisters/tojserverv2/domain/approve/policy/ApproveCountPolicy.kt
@@ -1,7 +1,7 @@
 package com.sexysisters.tojserverv2.domain.approve.policy
 
-import com.sexysisters.tojserverv2.domain.student.Student
-import com.sexysisters.tojserverv2.domain.student.engaged
+import com.sexysisters.tojserverv2.domain.student.domain.Student
+import com.sexysisters.tojserverv2.domain.student.domain.engaged
 import org.springframework.core.annotation.Order
 import org.springframework.stereotype.Component
