+++ src/main/kotlin/com/sexysisters/tojserverv2/domain/approve/policy/ApprovePolicy.kt
@@ -1,6 +1,6 @@
 package com.sexysisters.tojserverv2.domain.approve.policy
 
-import com.sexysisters.tojserverv2.domain.student.Student
+import com.sexysisters.tojserverv2.domain.student.domain.Student
 
 interface ApprovePolicy {
     fun check(applicant: Student, acceptor: Student)
