+++ src/main/kotlin/com/sexysisters/tojserverv2/domain/approve/policy/SameSchoolPolicy.kt
@@ -1,7 +1,7 @@
 package com.sexysisters.tojserverv2.domain.approve.policy
 
 import com.sexysisters.tojserverv2.domain.approve.exception.ApproveException
-import com.sexysisters.tojserverv2.domain.student.Student
+import com.sexysisters.tojserverv2.domain.student.domain.Student
 import org.springframework.core.annotation.Order
 import org.springframework.stereotype.Component
