+++ src/main/kotlin/com/sexysisters/tojserverv2/domain/school/School.kt
@@ -4,7 +4,7 @@ import com.sexysisters.tojserverv2.domain.BaseTimeEntity
 import com.sexysisters.tojserverv2.domain.school.exception.SchoolException
 import com.sexysisters.tojserverv2.domain.school.type.Division
 import com.sexysisters.tojserverv2.domain.school.type.Kind
-import com.sexysisters.tojserverv2.domain.student.Student
+import com.sexysisters.tojserverv2.domain.student.domain.Student
 import com.sexysisters.tojserverv2.domain.teacher.Teacher
 import com.sexysisters.tojserverv2.domain.wiki.Wiki
 import javax.persistence.*
