+++ src/main/kotlin/com/sexysisters/tojserverv2/domain/school/policy/AlreadyExistsStudentPolicy.kt
@@ -2,7 +2,7 @@ package com.sexysisters.tojserverv2.domain.school.policy
 
 import com.sexysisters.tojserverv2.domain.school.School
 import com.sexysisters.tojserverv2.domain.school.exception.SchoolException
-import com.sexysisters.tojserverv2.domain.student.Student
+import com.sexysisters.tojserverv2.domain.student.domain.Student
 import com.sexysisters.tojserverv2.infrastructure.student.StudentRepository
 import org.springframework.core.annotation.Order
 import org.springframework.stereotype.Component
