+++ src/main/kotlin/com/sexysisters/tojserverv2/domain/school/policy/DirectJoinPolicy.kt
@@ -1,9 +1,9 @@
 package com.sexysisters.tojserverv2.domain.school.policy
 
 import com.sexysisters.tojserverv2.domain.school.School
-import com.sexysisters.tojserverv2.domain.student.Student
-import com.sexysisters.tojserverv2.domain.student.engaged
-import com.sexysisters.tojserverv2.domain.student.waiting
+import com.sexysisters.tojserverv2.domain.student.domain.Student
+import com.sexysisters.tojserverv2.domain.student.domain.engaged
+import com.sexysisters.tojserverv2.domain.student.domain.waiting
 import org.springframework.core.annotation.Order
 import org.springframework.stereotype.Component
