+++ src/main/kotlin/com/sexysisters/tojserverv2/domain/school/policy/SchoolPolicy.kt
@@ -1,7 +1,7 @@
 package com.sexysisters.tojserverv2.domain.school.policy
 
 import com.sexysisters.tojserverv2.domain.school.School
-import com.sexysisters.tojserverv2.domain.student.Student
+import com.sexysisters.tojserverv2.domain.student.domain.Student
 
 interface SchoolPolicy {
     fun check(student: Student, school: School)
