+++ src/main/kotlin/com/sexysisters/tojserverv2/domain/student/StudentMapper.kt
@@ -1,5 +1,6 @@
 package com.sexysisters.tojserverv2.domain.student
 
+import com.sexysisters.tojserverv2.domain.student.domain.Student
 import org.mapstruct.InjectionStrategy
 import org.mapstruct.Mapper
 import org.mapstruct.Mapping
