+++ src/main/kotlin/com/sexysisters/tojserverv2/domain/student/StudentReader.kt
@@ -1,6 +1,8 @@
 package com.sexysisters.tojserverv2.domain.student
 
 import com.sexysisters.tojserverv2.domain.school.School
+import com.sexysisters.tojserverv2.domain.student.domain.Status
+import com.sexysisters.tojserverv2.domain.student.domain.Student
 
 interface StudentReader {
     fun getStudent(id: Long): Student
