+++ src/main/kotlin/com/sexysisters/tojserverv2/domain/student/StudentStore.kt
@@ -1,5 +1,7 @@
 package com.sexysisters.tojserverv2.domain.student
 
+import com.sexysisters.tojserverv2.domain.student.domain.Student
+
 interface StudentStore {
     fun store(student: Student): Long
 }
